+++ core/modules/logger/index.js
@@ -1,1 +1,2 @@
-export { createLogger, createRequestLogger } from './src/logger.js';
+export { createLogger } from './src/logger.js';
+export { createRequestLogger } from './src/request-logger.js';